+++ public/pages/EditRollup/containers/EditRollup.test.tsx
@@ -26,7 +26,7 @@ import { ServicesConsumer, ServicesContext } from "../../../services";
 import { ModalProvider, ModalRoot } from "../../../components/Modal";
 import { BrowserServices } from "../../../models/interfaces";
 import { BREADCRUMBS, ROUTES } from "../../../utils/constants";
-import { testRollup } from "../../../utils/constants.test";
+import { testRollup } from "../../../utils/testConstants";
 
 function renderEditRollupWithRouter(initialEntries = ["/"]) {
   return {
